requirements.txt
aiohttp
asyncpg
aioredis
pydantic
transformers
torch  
kafka-python
streamlit
pymilvus
nest_asyncio